debug

+++ data_loader.py
@@ -73,9 +73,9 @@ class MnistLoader(object):
         # im.save("demo.png")
 
     def pca(self, var_per=0.99):
-        lowD_train = pca(self.data_train, var_per)
+        lowD_train = pca(self.data_train, var_per).astype(np.float32)
         pca_op = PCA(n_components=lowD_train.shape[1])
-        lowD_test = pca_op.fit_transform(self.data_test)
+        lowD_test = pca_op.fit_transform(self.data_test).astype(np.float32)
 
         print('training set shape: ', lowD_train.shape)
         print('test set shape: ', lowD_test.shape)
